remove Informations component
